Revert "refactor(audit): isolate row-HMAC signing primitives in audit-hmac.ts"
This reverts commit 17ae667a.

The extraction moved the (byte-identical) computeRowHmacV1/V2 functions to a new
file, which relocated a pre-existing CodeQL js/insufficient-password-hash finding
to a new path — making it a 'new' PR alert even though the createHmac("sha256")
integrity-signing code is unchanged from main. The finding is a false positive
(HMAC row-tamper-evidence, not password storage; SHA-256 HMAC is the correct
primitive and cannot change without invalidating every signed audit row).

Rather than add an unapproved security-scan suppression for the lowest-value
commit in this refactor, drop the extraction. audit.ts returns to its exact
prior state; the higher-value build.ts and user-placeholder extractions stand.
